fixed mapping

+++ src/main/java/com/selina/lending/internal/dto/DetailDto.java
@@ -58,6 +58,6 @@ public class DetailDto {
     String date;
     Double currentBalance;
     String interestOnlyBalance;
-    Double amount;
+    Integer amount;
     String code;
 }
